First Pass at Block Collision
This works on blocks falling from above. It does not work for blocks being placed below overhangs, and also doesn't prevent blocks from clipping through other blocks from the side.
The side collision could be handled in a similar way to how the frame collision is handled in the move() function, but I need to refactor this. It is getting complicated quickly and needs to be cleaned up.