WORKING COLLISION!
The last collision methods did not work with overhangs or with colliding pieces from the side.

This new collision method works with all of these cases and also allows for easy line checking and removal.

I still need to fix the bug allowing for clipping when you rotate near structures.

Also, the new functions were put in quickly, my next commit should be to encapsulate those in either the piece class or a new GameBoard class. Using a GameBoard class will allow for easy multiplayer later on.