Start and Restart Adjustments
The First game now prompts you to press space to begin.
When a gameover is reached, the game now properly starts again once you press space. Before there was a bug where paused was never set back to true on a gameover state, so when you would press space, the game would reset the scoreboard, but never update and reset the gameboard.