Kill functions
Added kill functions which are called on the previous classes stored in GBArr when the GameState is changed.

These functions ensure that their update methods are stopped, and that they are fully removed from the DOM before continuing to the next GameState.

All future classes that will be called directly from the GSH must have a kill method.