SplashScreen Update!
The game now has a working SplashScreen. It still needs to be trimmed down to the essentials and made more efficient, but it will work for now.

When you press space, the game currently moves to the single player mode. This is fine for now, but my next move needs to be a mode selection screen so I can choose between multiplayer and singleplayer.